Component audit - Modify/Delete API OS tests (#11826)
* follow ohsnap repofile redirects

* component audit cli operating system

---------

(cherry picked from commit 531f31b11836cfda6a2f8dda99863675880bd2fa)
